+++ src/types.ts
@@ -56,10 +56,12 @@ export type DeclarationType = 'api' | 'none' | 'standard';
 
 // PACKAGES
 
+export type InputMap = Record<string, string>;
+
 export interface PackemonPackageConfig {
   bundle?: boolean;
   format?: Format | Format[];
-  inputs?: Record<string, string>;
+  inputs?: InputMap;
   namespace?: string;
   platform?: Platform | Platform[];
   support?: Support;
@@ -72,7 +74,7 @@ export interface PackemonPackage extends PackageStructure {
 export interface PackageConfig {
   bundle: boolean;
   formats: Format[];
-  inputs: Record<string, string>;
+  inputs: InputMap;
   namespace: string;
   platform: Platform;
   support: Support;
@@ -127,6 +129,11 @@ export interface BundleBuild {
   format: Format;
   platform: Platform;
   support: Support;
+  stats?: { size: number };
+}
+
+export interface CodeBuild {
+  format: Format;
   stats?: { size: number };
 }
